refactor: Repository 구조 단순화 (entity.Repository 삭제) - 댓글
